+++ build.sbt
@@ -134,9 +134,12 @@ lazy val scalaMacroDependencies: Seq[Setting[_]] = Seq(
 )
 
 lazy val examplesDependencies: Seq[Setting[_]] = Seq(
-  libraryDependencies += "org.typelevel" %% "cats-core" % "0.9.0",
-  libraryDependencies += "com.propensive" %% "estrapade" % "1.0.0",
-  libraryDependencies += "com.propensive" %% "contextual-data" % "1.0.3"
+  libraryDependencies ++= Seq(
+    "com.fommil" %% "stalactite" % "0.0.5",
+    "org.typelevel" %% "cats-core" % "0.9.0"
+      //"com.propensive" %% "estrapade" % "1.0.0",
+      //"com.propensive" %% "contextual-data" % "1.0.3"
+  )
 )
 
 credentials ++= (for {
